2.0.4 - tweak server live reloading

+++ app/templates/gulp/tasks/serve.js
@@ -1,37 +1,38 @@
 import gulp from 'gulp';
 // import config from '../config';
-import liveReload from 'gulp-livereload';
 import { log, colors } from 'gulp-util';
 import { spawn } from 'child_process';
 
 let currentProcess = null;
+let bootingProcess = null;
 
-function checkForBoot(chunk) {
+function checkForBoot(proc, chunk) {
     const chunkAsString = chunk.toString();
     if (/Magic/.test(chunkAsString)) {
-        currentProcess.removeListener('data', checkForBoot);
+        bootingProcess = null;
+        if (currentProcess) currentProcess.kill();
+        currentProcess = proc;
+        proc.removeListener('data', checkForBoot);
         log(colors.green('Server up'));
-        liveReload.reload();
+        // liveReload.reload();
     }
 }
-function handleClose(code) {
+function handleClose(proc, code) {
     if (code === 8) {
         log('Error detected, waiting for changes...');
     } else {
         log('node closed: ', code);
     }
-    currentProcess.removeListener('close', handleClose);
+    proc.removeListener('close', handleClose);
 }
 
 gulp.task('serve', function() {
-    if (currentProcess) {
-        currentProcess.kill();
-    }
-    currentProcess = spawn('node', ['server/dev-server']);
-    currentProcess.on('close', handleClose);
-    currentProcess.stdout.on('data', checkForBoot);
-    currentProcess.stdout.pipe(process.stdout);
-    currentProcess.stderr.pipe(process.stderr);
+    if (bootingProcess) bootingProcess.kill();
+    const nextProcess = bootingProcess = spawn('node', ['server/dev-server']);
+    nextProcess.on('close', handleClose.bind(null, nextProcess));
+    nextProcess.stdout.on('data', checkForBoot.bind(null, nextProcess));
+    nextProcess.stdout.pipe(process.stdout);
+    nextProcess.stderr.pipe(process.stderr);
 });
 
 // clean up if an error goes unhandled.
